Split manifold test
