+++ test/manifold_advanced_test.cpp
@@ -1,0 +1,417 @@
+// Copyright 2021 The Manifold Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include <algorithm>
+
+#include "cross_section.h"
+#include "manifold.h"
+#include "samples.h"
+#include "test.h"
+#include "tri_dist.h"
+
+namespace {
+
+using namespace manifold;
+
+template <typename T>
+int NumUnique(const std::vector<T>& in) {
+  std::set<int> unique;
+  for (const T& v : in) {
+    unique.emplace(v);
+  }
+  return unique.size();
+}
+
+}  // namespace
+
+/**
+ * Testing more advanced Manifold operations.
+ */
+
+TEST(Manifold, Transform) {
+  Manifold cube = Manifold::Cube({1, 2, 3});
+  Manifold cube2 = cube;
+  cube = cube.Rotate(30, 40, 50).Scale({6, 5, 4}).Translate({1, 2, 3});
+
+  glm::mat3 rX(1.0f, 0.0f, 0.0f,          //
+               0.0f, cosd(30), sind(30),  //
+               0.0f, -sind(30), cosd(30));
+  glm::mat3 rY(cosd(40), 0.0f, -sind(40),  //
+               0.0f, 1.0f, 0.0f,           //
+               sind(40), 0.0f, cosd(40));
+  glm::mat3 rZ(cosd(50), sind(50), 0.0f,   //
+               -sind(50), cosd(50), 0.0f,  //
+               0.0f, 0.0f, 1.0f);
+  glm::mat3 s = glm::mat3(1.0f);
+  s[0][0] = 6;
+  s[1][1] = 5;
+  s[2][2] = 4;
+  glm::mat4x3 transform = glm::mat4x3(s * rZ * rY * rX);
+  transform[3] = glm::vec3(1, 2, 3);
+  cube2 = cube2.Transform(transform);
+
+  Identical(cube.GetMesh(), cube2.GetMesh());
+}
+
+TEST(Manifold, Slice) {
+  Manifold cube = Manifold::Cube();
+  CrossSection bottom = cube.Slice();
+  CrossSection top = cube.Slice(1);
+  EXPECT_EQ(bottom.Area(), 1);
+  EXPECT_EQ(top.Area(), 0);
+}
+
+TEST(Manifold, MeshRelation) {
+  Mesh gyroidMesh = Gyroid();
+  MeshGL gyroidMeshGL = WithIndexColors(gyroidMesh);
+  Manifold gyroid(gyroidMeshGL);
+
+#ifdef MANIFOLD_EXPORT
+  ExportOptions opt;
+  opt.mat.roughness = 1;
+  opt.mat.colorChannels = glm::ivec4(3, 4, 5, -1);
+  if (options.exportModels) ExportMesh("gyroid.glb", gyroid.GetMeshGL(), opt);
+#endif
+
+  RelatedGL(gyroid, {gyroidMeshGL});
+}
+
+TEST(Manifold, MeshRelationTransform) {
+  const Manifold cube = Manifold::Cube();
+  const MeshGL cubeGL = cube.GetMeshGL();
+  const Manifold turned = cube.Rotate(45, 90);
+
+  RelatedGL(turned, {cubeGL});
+}
+
+TEST(Manifold, MeshRelationRefine) {
+  const Mesh in = Csaszar();
+  MeshGL inGL = WithIndexColors(in);
+  Manifold csaszar(inGL);
+
+  RelatedGL(csaszar, {inGL});
+  csaszar = csaszar.RefineToLength(1);
+  ExpectMeshes(csaszar, {{9019, 18038, 3}});
+  RelatedGL(csaszar, {inGL});
+
+#ifdef MANIFOLD_EXPORT
+  ExportOptions opt;
+  opt.mat.roughness = 1;
+  opt.mat.colorChannels = glm::ivec4(3, 4, 5, -1);
+  if (options.exportModels) ExportMesh("csaszar.glb", csaszar.GetMeshGL(), opt);
+#endif
+}
+
+TEST(Manifold, MeshGLRoundTrip) {
+  const Manifold cylinder = Manifold::Cylinder(2, 1);
+  ASSERT_GE(cylinder.OriginalID(), 0);
+  MeshGL inGL = cylinder.GetMeshGL();
+  const Manifold cylinder2(inGL);
+  const MeshGL outGL = cylinder2.GetMeshGL();
+
+  ASSERT_EQ(inGL.runOriginalID.size(), 1);
+  ASSERT_EQ(outGL.runOriginalID.size(), 1);
+  ASSERT_EQ(outGL.runOriginalID[0], inGL.runOriginalID[0]);
+
+  RelatedGL(cylinder2, {inGL});
+}
+
+void CheckCube(const MeshGL& cubeSTL) {
+  Manifold cube(cubeSTL);
+  EXPECT_EQ(cube.NumTri(), 12);
+  EXPECT_EQ(cube.NumVert(), 8);
+  EXPECT_EQ(cube.NumPropVert(), 24);
+
+  auto prop = cube.GetProperties();
+  EXPECT_FLOAT_EQ(prop.volume, 1.0f);
+  EXPECT_FLOAT_EQ(prop.surfaceArea, 6.0f);
+}
+
+TEST(Manifold, Merge) {
+  MeshGL cubeSTL = CubeSTL();
+  EXPECT_EQ(cubeSTL.NumTri(), 12);
+  EXPECT_EQ(cubeSTL.NumVert(), 36);
+
+  Manifold cubeBad(cubeSTL);
+  EXPECT_TRUE(cubeBad.IsEmpty());
+  EXPECT_EQ(cubeBad.Status(), Manifold::Error::NotManifold);
+
+  EXPECT_TRUE(cubeSTL.Merge());
+  CheckCube(cubeSTL);
+
+  EXPECT_FALSE(cubeSTL.Merge());
+  EXPECT_EQ(cubeSTL.mergeFromVert.size(), 28);
+  cubeSTL.mergeFromVert.resize(14);
+  cubeSTL.mergeToVert.resize(14);
+
+  EXPECT_TRUE(cubeSTL.Merge());
+  EXPECT_EQ(cubeSTL.mergeFromVert.size(), 28);
+  CheckCube(cubeSTL);
+}
+
+TEST(Manifold, FaceIDRoundTrip) {
+  const Manifold cube = Manifold::Cube();
+  ASSERT_GE(cube.OriginalID(), 0);
+  MeshGL inGL = cube.GetMeshGL();
+  ASSERT_EQ(NumUnique(inGL.faceID), 6);
+  inGL.faceID = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11};
+
+  const Manifold cube2(inGL);
+  const MeshGL outGL = cube2.GetMeshGL();
+  ASSERT_EQ(NumUnique(outGL.faceID), 12);
+}
+
+TEST(Manifold, MirrorUnion) {
+  auto a = Manifold::Cube({5., 5., 5.}, true);
+  auto b = a.Translate({2.5, 2.5, 2.5});
+  auto result = a + b + b.Mirror({1, 1, 0});
+
+#ifdef MANIFOLD_EXPORT
+  if (options.exportModels)
+    ExportMesh("manifold_mirror_union.glb", result.GetMesh(), {});
+#endif
+
+  auto vol_a = a.GetProperties().volume;
+  EXPECT_FLOAT_EQ(vol_a * 2.75, result.GetProperties().volume);
+  EXPECT_TRUE(a.Mirror(glm::vec3(0)).IsEmpty());
+}
+
+TEST(Manifold, MirrorUnion2) {
+  auto a = Manifold::Cube();
+  auto result = Manifold::Compose({a.Mirror({1, 0, 0})});
+  EXPECT_TRUE(result.MatchesTriNormals());
+}
+
+TEST(Manifold, Invalid) {
+  auto invalid = Manifold::Error::InvalidConstruction;
+  auto circ = CrossSection::Circle(10.);
+  auto empty_circ = CrossSection::Circle(-2.);
+  auto empty_sq = CrossSection::Square(glm::vec2(0.0f));
+
+  EXPECT_EQ(Manifold::Sphere(0).Status(), invalid);
+  EXPECT_EQ(Manifold::Cylinder(0, 5).Status(), invalid);
+  EXPECT_EQ(Manifold::Cylinder(2, -5).Status(), invalid);
+  EXPECT_EQ(Manifold::Cube(glm::vec3(0.0f)).Status(), invalid);
+  EXPECT_EQ(Manifold::Cube({-1, 1, 1}).Status(), invalid);
+  EXPECT_EQ(Manifold::Extrude(circ, 0.).Status(), invalid);
+  EXPECT_EQ(Manifold::Extrude(empty_circ, 10.).Status(), invalid);
+  EXPECT_EQ(Manifold::Revolve(empty_sq).Status(), invalid);
+}
+
+TEST(Manifold, MultiCompose) {
+  auto part = Manifold::Compose({Manifold::Cube({10, 10, 10})});
+  auto finalAssembly =
+      Manifold::Compose({part, part.Translate({0, 10, 0}),
+                         part.Mirror({1, 0, 0}).Translate({10, 0, 0}),
+                         part.Mirror({1, 0, 0}).Translate({10, 10, 0})});
+  EXPECT_FLOAT_EQ(finalAssembly.GetProperties().volume, 4000);
+}
+
+TEST(Manifold, MergeDegenerates) {
+  MeshGL cube = Manifold::Cube(glm::vec3(1), true).GetMeshGL();
+  MeshGL squash;
+  squash.vertProperties = cube.vertProperties;
+  squash.triVerts = cube.triVerts;
+  // Move one vert to the position of its neighbor and remove one triangle
+  // linking them to break the manifold.
+  squash.vertProperties[squash.vertProperties.size() - 1] *= -1;
+  squash.triVerts.resize(squash.triVerts.size() - 3);
+  // Rotate the degenerate triangle to the middle to catch more problems.
+  std::rotate(squash.triVerts.begin(), squash.triVerts.begin() + 3 * 5,
+              squash.triVerts.end());
+  // Merge should remove the now duplicate vertex.
+  EXPECT_TRUE(squash.Merge());
+  // Manifold should remove the triangle with two references to the same vert.
+  Manifold squashed = Manifold(squash);
+  EXPECT_FALSE(squashed.IsEmpty());
+  EXPECT_EQ(squashed.Status(), Manifold::Error::NoError);
+}
+
+TEST(Manifold, PinchedVert) {
+  Mesh shape;
+  shape.vertPos = {{0, 0, 0},         //
+                   {1, 1, 0},         //
+                   {1, -1, 0},        //
+                   {-0.00001, 0, 0},  //
+                   {-1, -1, -0},      //
+                   {-1, 1, 0},        //
+                   {0, 0, 2},         //
+                   {0, 0, -2}};
+  shape.triVerts = {{0, 2, 6},  //
+                    {2, 1, 6},  //
+                    {1, 0, 6},  //
+                    {4, 3, 6},  //
+                    {3, 5, 6},  //
+                    {5, 4, 6},  //
+                    {2, 0, 4},  //
+                    {0, 3, 4},  //
+                    {3, 0, 1},  //
+                    {3, 1, 5},  //
+                    {7, 2, 4},  //
+                    {7, 4, 5},  //
+                    {7, 5, 1},  //
+                    {7, 1, 2}};
+  Manifold touch(shape);
+  EXPECT_FALSE(touch.IsEmpty());
+  EXPECT_EQ(touch.Status(), Manifold::Error::NoError);
+  EXPECT_EQ(touch.Genus(), 0);
+}
+
+TEST(Manifold, TictacHull) {
+  const float tictacRad = 100;
+  const float tictacHeight = 500;
+  const int tictacSeg = 1000;
+  const float tictacMid = tictacHeight - 2 * tictacRad;
+  const auto sphere = Manifold::Sphere(tictacRad, tictacSeg);
+  const std::vector<Manifold> spheres{sphere,
+                                      sphere.Translate({0, 0, tictacMid})};
+  const auto tictac = Manifold::Hull(spheres);
+
+#ifdef MANIFOLD_EXPORT
+  if (options.exportModels) {
+    ExportMesh("tictac_hull.glb", tictac.GetMesh(), {});
+  }
+#endif
+
+  EXPECT_EQ(sphere.NumVert() + tictacSeg, tictac.NumVert());
+}
+
+#ifdef MANIFOLD_EXPORT
+TEST(Manifold, HullFail) {
+  Manifold body = ReadMesh("hull-body.glb");
+  Manifold mask = ReadMesh("hull-mask.glb");
+  Manifold ret = body - mask;
+  MeshGL mesh = ret.GetMesh();
+}
+#endif
+
+TEST(Manifold, HollowHull) {
+  auto sphere = Manifold::Sphere(100, 360);
+  auto hollow = sphere - sphere.Scale({0.8, 0.8, 0.8});
+  const float sphere_vol = sphere.GetProperties().volume;
+  EXPECT_FLOAT_EQ(hollow.Hull().GetProperties().volume, sphere_vol);
+}
+
+TEST(Manifold, CubeHull) {
+  std::vector<glm::vec3> cubePts = {
+      {0, 0, 0},       {1, 0, 0},   {0, 1, 0},      {0, 0, 1},  // corners
+      {1, 1, 0},       {0, 1, 1},   {1, 0, 1},      {1, 1, 1},  // corners
+      {0.5, 0.5, 0.5}, {0.5, 0, 0}, {0.5, 0.7, 0.2}  // internal points
+  };
+  auto cube = Manifold::Hull(cubePts);
+  EXPECT_FLOAT_EQ(cube.GetProperties().volume, 1);
+}
+
+TEST(Manifold, EmptyHull) {
+  const std::vector<glm::vec3> tooFew{{0, 0, 0}, {1, 0, 0}, {0, 1, 0}};
+  EXPECT_TRUE(Manifold::Hull(tooFew).IsEmpty());
+
+  const std::vector<glm::vec3> coplanar{
+      {0, 0, 0}, {1, 0, 0}, {0, 1, 0}, {1, 1, 0}};
+  EXPECT_TRUE(Manifold::Hull(coplanar).IsEmpty());
+}
+
+TEST(Manifold, TriangleDistanceClosestPointsOnVertices) {
+  std::array<glm::vec3, 3> p = {glm::vec3{-1, 0, 0}, glm::vec3{1, 0, 0},
+                                glm::vec3{0, 1, 0}};
+
+  std::array<glm::vec3, 3> q = {glm::vec3{2, 0, 0}, glm::vec3{4, 0, 0},
+                                glm::vec3{3, 1, 0}};
+
+  float distance = DistanceTriangleTriangleSquared(p, q);
+
+  EXPECT_FLOAT_EQ(distance, 1);
+}
+
+TEST(Manifold, TriangleDistanceClosestPointOnEdge) {
+  std::array<glm::vec3, 3> p = {glm::vec3{-1, 0, 0}, glm::vec3{1, 0, 0},
+                                glm::vec3{0, 1, 0}};
+
+  std::array<glm::vec3, 3> q = {glm::vec3{-1, 2, 0}, glm::vec3{1, 2, 0},
+                                glm::vec3{0, 3, 0}};
+
+  float distance = DistanceTriangleTriangleSquared(p, q);
+
+  EXPECT_FLOAT_EQ(distance, 1);
+}
+
+TEST(Manifold, TriangleDistanceClosestPointOnEdge2) {
+  std::array<glm::vec3, 3> p = {glm::vec3{-1, 0, 0}, glm::vec3{1, 0, 0},
+                                glm::vec3{0, 1, 0}};
+
+  std::array<glm::vec3, 3> q = {glm::vec3{1, 1, 0}, glm::vec3{3, 1, 0},
+                                glm::vec3{2, 2, 0}};
+
+  float distance = DistanceTriangleTriangleSquared(p, q);
+
+  EXPECT_FLOAT_EQ(distance, 0.5f);
+}
+
+TEST(Manifold, TriangleDistanceClosestPointOnFace) {
+  std::array<glm::vec3, 3> p = {glm::vec3{-1, 0, 0}, glm::vec3{1, 0, 0},
+                                glm::vec3{0, 1, 0}};
+
+  std::array<glm::vec3, 3> q = {glm::vec3{-1, 2, -0.5f}, glm::vec3{1, 2, -0.5f},
+                                glm::vec3{0, 2, 1.5f}};
+
+  float distance = DistanceTriangleTriangleSquared(p, q);
+
+  EXPECT_FLOAT_EQ(distance, 1);
+}
+
+TEST(Manifold, TriangleDistanceOverlapping) {
+  std::array<glm::vec3, 3> p = {glm::vec3{-1, 0, 0}, glm::vec3{1, 0, 0},
+                                glm::vec3{0, 1, 0}};
+
+  std::array<glm::vec3, 3> q = {glm::vec3{-1, 0, 0}, glm::vec3{1, 0.5f, 0},
+                                glm::vec3{0, 1, 0}};
+
+  float distance = DistanceTriangleTriangleSquared(p, q);
+
+  EXPECT_FLOAT_EQ(distance, 0);
+}
+
+/**
+ * Curvature is the inverse of the radius of curvature, and signed such that
+ * positive is convex and negative is concave. There are two orthogonal
+ * principal curvatures at any point on a manifold, with one maximum and the
+ * other minimum. Gaussian curvature is their product, while mean
+ * curvature is their sum. Here we check our discrete approximations calculated
+ * at each vertex against the constant expected values of spheres of different
+ * radii and at different mesh resolutions.
+ */
+TEST(Manifold, CalculateCurvature) {
+  const float precision = 0.015;
+  for (int n = 4; n < 100; n *= 2) {
+    const int gaussianIdx = 3;
+    const int meanIdx = 4;
+    Manifold sphere = Manifold::Sphere(1, 64).CalculateCurvature(
+        gaussianIdx - 3, meanIdx - 3);
+    MeshGL sphereGL = sphere.GetMeshGL();
+    ASSERT_EQ(sphereGL.numProp, 5);
+    EXPECT_NEAR(GetMinProperty(sphereGL, meanIdx), 2, 2 * precision);
+    EXPECT_NEAR(GetMaxProperty(sphereGL, meanIdx), 2, 2 * precision);
+    EXPECT_NEAR(GetMinProperty(sphereGL, gaussianIdx), 1, precision);
+    EXPECT_NEAR(GetMaxProperty(sphereGL, gaussianIdx), 1, precision);
+
+    sphere = sphere.Scale(glm::vec3(2.0f))
+                 .CalculateCurvature(gaussianIdx - 3, meanIdx - 3);
+    sphereGL = sphere.GetMeshGL();
+    ASSERT_EQ(sphereGL.numProp, 5);
+    EXPECT_NEAR(GetMinProperty(sphereGL, meanIdx), 1, precision);
+    EXPECT_NEAR(GetMaxProperty(sphereGL, meanIdx), 1, precision);
+    EXPECT_NEAR(GetMinProperty(sphereGL, gaussianIdx), 0.25, 0.25 * precision);
+    EXPECT_NEAR(GetMaxProperty(sphereGL, gaussianIdx), 0.25, 0.25 * precision);
+  }
+}
